backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
loguru
python-dotenv
neo4j
qdrant-client
pandas
numpy
networkx
pdfplumber
python-docx
chardet
